T: inline commenter tests
