+++ src/test/kotlin/org/rust/ide/commenter/RsCommenterTest.kt
@@ -6,29 +6,126 @@
 package org.rust.ide.commenter
 
 import com.intellij.openapi.actionSystem.IdeActions
+import org.intellij.lang.annotations.Language
 import org.rust.RsTestBase
 
 class RsCommenterTest : RsTestBase() {
 
-    override val dataPath = "org/rust/ide/commenter/fixtures"
+    fun `test single line`() = checkEditorAction("""
+        fn <caret>double(x: i32) -> i32 {
+            x * 2
+        }
+    """, """
+        //fn double(x: i32) -> i32 {
+            x<caret> * 2
+        }
+    """, IdeActions.ACTION_COMMENT_LINE)
 
-    private fun doTest(actionId: String) {
-        myFixture.configureByFile(fileName)
-        myFixture.performEditorAction(actionId)
-        myFixture.checkResultByFile(fileName.replace(".rs", "_after.rs"), true)
+    fun `test multi line`() = checkEditorAction("""
+        fn doub<selection>le(x: i32) -> i32 {
+            x</selection> * 2
+        }
+    """, """
+        //fn doub<selection>le(x: i32) -> i32 {
+        //    x</selection> * 2
+        }
+    """, IdeActions.ACTION_COMMENT_LINE)
+
+    fun `test single line block`() = checkEditorAction("""
+        fn d<caret>ouble(x: i32) -> i32 {
+            x * 2
+        }
+    """, """
+        fn d/*<caret>*/ouble(x: i32) -> i32 {
+            x * 2
+        }
+    """, IdeActions.ACTION_COMMENT_BLOCK)
+
+    fun `test multi line block`() = checkEditorAction("""
+        fn doub<selection>le(x: i32) -> i32 {
+            x</selection> * 2
+        }
+    """, """
+        fn doub<selection>/*le(x: i32) -> i32 {
+            x*/</selection> * 2
+        }
+    """, IdeActions.ACTION_COMMENT_BLOCK)
+
+    fun `test single line uncomment`() = checkEditorAction("""
+        //fn d<caret>ouble(x: i32) -> i32 {
+        //    x * 2
+        }
+    """, """
+        fn double(x: i32) -> i32 {
+        //  <caret>  x * 2
+        }
+    """, IdeActions.ACTION_COMMENT_LINE)
+
+    fun `test multi line uncomment`() = checkEditorAction("""
+        //fn doub<selection>le(x: i32) -> i32 {
+        //    x</selection> * 2
+        }
+    """, """
+        fn doub<selection>le(x: i32) -> i32 {
+            x</selection> * 2
+        }
+    """, IdeActions.ACTION_COMMENT_LINE)
+
+    fun `test single line block uncomment`() = checkEditorAction("""
+        fn d/*<caret>*/ouble(x: i32) -> i32 {
+            x * 2
+        }
+    """, """
+        fn d<caret>ouble(x: i32) -> i32 {
+            x * 2
+        }
+    """, IdeActions.ACTION_COMMENT_BLOCK)
+
+    fun `test multi line block uncomment`() = checkEditorAction("""
+        fn doub<selection>/*le(x: i32) -> i32 {
+            x*/</selection> * 2
+        }
+    """, """
+        fn doub<selection>le(x: i32) -> i32 {
+            x</selection> * 2
+        }
+    """, IdeActions.ACTION_COMMENT_BLOCK)
+
+    fun `test single line uncomment with space`() = checkEditorAction("""
+        // fn d<caret>ouble(x: i32) -> i32 {
+        //     x * 2
+        }
+    """, """
+         fn double(x: i32) -> i32 {
+        //   <caret>  x * 2
+        }
+    """, IdeActions.ACTION_COMMENT_LINE)
+
+    fun `test nested block comments`() = checkEditorAction("""
+        fn double<selection>(x: i32/* foobar */) -> i32</selection> {
+            x * 2
+        }
+    """, """
+        fn double<selection>/*(x: i32*//* foobar *//*) -> i32*/</selection> {
+            x * 2
+        }
+    """, IdeActions.ACTION_COMMENT_BLOCK) // FIXME
+
+    fun `test indented single line comment`() = checkEditorAction("""
+        fn double(x: i32) -> i32 {
+            x<caret> * 2
+        }
+    """, """
+        fn double(x: i32) -> i32 {
+        //    x * 2
+        }<caret>
+    """, IdeActions.ACTION_COMMENT_LINE)
+
+    private fun checkEditorAction(
+        @Language("Rust") before: String,
+        @Language("Rust") after: String,
+        actionId: String
+    ) {
+        checkByText(before.trimIndent(), after.trimIndent()) { myFixture.performEditorAction(actionId) }
     }
-
-    fun `test single line`() = doTest(IdeActions.ACTION_COMMENT_LINE)
-    fun `test multi line`() = doTest(IdeActions.ACTION_COMMENT_LINE)
-    fun `test single line block`() = doTest(IdeActions.ACTION_COMMENT_BLOCK)
-    fun `test multi line block`() = doTest(IdeActions.ACTION_COMMENT_BLOCK)
-
-    fun `test single line uncomment`() = doTest(IdeActions.ACTION_COMMENT_LINE)
-    fun `test multi line uncomment`() = doTest(IdeActions.ACTION_COMMENT_LINE)
-    fun `test single line block uncomment`() = doTest(IdeActions.ACTION_COMMENT_BLOCK)
-    fun `test multi line block uncomment`() = doTest(IdeActions.ACTION_COMMENT_BLOCK)
-
-    fun `test single line uncomment with space`() = doTest(IdeActions.ACTION_COMMENT_LINE)
-    fun `test nested block comments`() = doTest(IdeActions.ACTION_COMMENT_BLOCK) // FIXME
-    fun `test indented single line comment`() = doTest(IdeActions.ACTION_COMMENT_LINE)
 }
