--- src/test/resources/org/rust/ide/commenter/fixtures/indented_single_line_comment_after.rs
@@ -1,3 +1,0 @@
-fn double(x: i32) -> i32 {
-//    x * 2
-}<caret>
