--- src/test/resources/org/rust/ide/commenter/fixtures/multi_line_after.rs
@@ -1,3 +1,0 @@
-//fn doub<selection>le(x: i32) -> i32 {
-//    x</selection> * 2
-}
