--- src/test/resources/org/rust/ide/commenter/fixtures/nested_block_comments_after.rs
@@ -1,3 +1,0 @@
-fn double<selection>/*(x: i32*//* foobar *//*) -> i32*/</selection> {
-    x * 2
-}
