--- src/test/resources/org/rust/ide/commenter/fixtures/single_line.rs
@@ -1,3 +1,0 @@
-fn d<caret>ouble(x: i32) -> i32 {
-    x * 2
-}
