--- src/test/resources/org/rust/ide/commenter/fixtures/single_line_uncomment_after.rs
@@ -1,3 +1,0 @@
-fn double(x: i32) -> i32 {
-//  <caret>  x * 2
-}
